chore: 分离结果组件
